+++ braidpool-primitives/src/bead/mod.rs
@@ -83,6 +83,4 @@ impl Bead {
 }
 
 #[cfg(test)]
-mod tests {
-    // Unit Tests for Private Functions
-}
+mod tests;
