+++ braidpool-primitives/src/lib.rs
@@ -1,3 +1,3 @@
-pub mod beads;
+pub mod bead;
 pub mod braid;
 pub mod utils;
