+++ braidpool-primitives/src/utils/bitcoin/mod.rs
@@ -83,3 +83,6 @@ impl From<MiningBlockHeader> for BlockHeader {
         }
     }
 }
+
+#[cfg(test)]
+mod tests;
